+++ src/Command/ImportCommand.php
@@ -54,7 +54,6 @@ class ImportCommand extends Command
 
         try {
             $file = (new BackupImport)->filename($args->getArgument('filename'))->import();
-
             $io->success(__d('database_backup', 'Backup `{0}` has been imported', rtr($file)));
         } catch (Exception $e) {
             $io->error($e->getMessage());