+++ src/Command/SendCommand.php
@@ -59,7 +59,6 @@ class SendCommand extends Command
 
         try {
             (new BackupManager)->send($args->getArgument('filename'), $args->getArgument('recipient'));
-
             $io->success(__d('database_backup', 'Backup `{0}` was sent via mail', rtr($args->getArgument('filename'))));
         } catch (Exception $e) {
             $io->error($e->getMessage());